Update patch set 1
Patch Set 1:

Hi, thanks for the patch. I've confirmed that you've submitted a License Agreement, so welcome to PDFium.  You can edit the /AUTHORS file if you like to include your name in the individuals section, keeping the list alphabetized, and include it in the next patch set you submit for this CL.

During the review, we may ask you to make small changes to the code, in which case you should re-upload a new patch set containing the changes, or explain why the suggested changes are not appropriate.

Patch-set: 1
Reviewer: Gerrit User 5495 <5495@fca2d64a-0420-3532-b200-12848c95a47b>