[SkiaPaths] Fixed missing texts.
In 3bigpreview.pdf, a lot of texts were not displayed because
|SkisState::m_rsxform| never gets reset to empty after drawing text
objects with RSX(rotation+scale matrix).

In DrawText(), after drawing a text object with RSX, if the next text
object has no RSX, |hasRSX| will contain the updated value as false,
and m_rsxform.isEmpty() will return true due to |SkisState::m_rsxform|
not getting reset, which causes FlushText() to be skipped. To fix it,
need to reset |SkisState::m_rsxform| when start drawing a new text
object.

Bug:pdfium:1355
Change-Id: I7476587d014a58467822a4dd5b1935760dabc924