[SkiaPaths] Fixed missing texts.
In 3bigpreview.pdf, a lot of texts were not displayed because
|m_rscform| never gets reset to empty after drawing RSX text objects.

In DrawText(), after drawing an RSX object, if the next text object has
no RSX, |hasRSX| will contain the updated value as false, and
m_rsxform.isEmpty() will return true due to |m_rsxform| not getting
reset, which causes FlushText() to be skipped. To fix it, need to
reset |m_rsxform| when start drawing a new text object.

Bug:pdfium:1355
Change-Id: I7476587d014a58467822a4dd5b1935760dabc924