[SkiaPaths]Vertical texts were drawn in reversed direction.
In ChS_simple.pdf, the Chinese characters are displayed vertically and
when characters are displayed in order, the sentence direction was
going upwards instead of downwards. This is because when the main font
indicates vertical display("Identity-V"), its descendant font should
inherit the vertical setup since there is only one way to display the
text object. LoadSubstFont() is the only method to sync the descendant
font's |m_bVertical| with the main font, but it was skipped in
CPDF_CIDFont::Load() due to the main font is embedded.

To fix this issue, need to add a method to sync the descendant font's
|m_bVertical| value with the main font if it indicates vertical
display(not default value), regardless whether the main font is
embedded or not.

Bug:pdfium:1376
Change-Id: I7e935b59745b063b82ec709afd847c6bc289937d