Introduce IPWL_SystemHandler and fxcrt::TimerHandlerIface for layering.
One of a growing number of callback interfaces for the pwl
windowing layer which is unsurprising given how much functionality
from above this layer wants. Consolidation may be possible down
the road.

fxcrt::TimerHandlerIface is split off from the systemhander so
that fxjs/ can used the timer functions without gaining knowledge
of the pwl/ layer.

Change-Id: I1c075aac40a0c2842519f9c6f61e031042d7208c