Make CPWL_Timer simpler.
Remove CPWL_TimerHandler, which was an object that served two
purposes: as an interface to dispatch a timer event, and as
an interface to obtain a systemhandler. The second isn't really
required, and the first can be provided by a simple nested
virtual class.

Also remove the start/stop timer mechanism, timers can be cancelled
by merely deleting them.

Change-Id: I5fc1305c4e1b2693d39a74bea10b677263e9e12c