Make CPWL_Timer simpler.
Remove CPWL_TimerHandler, which was an object that served two
purposes: as an interface to dispatch a timer event, and as
an interface to obtain a systemhandler. The second isn't required,
and the first is better provided by a nested pure virtual class.

Also remove the start/stop timer mechanism, timers can be cancelled
by merely deleting them.

-- CPWL_Wnd now does not get a timer by default, only specific
   subclasses will have one.

Change-Id: I5fc1305c4e1b2693d39a74bea10b677263e9e12c