Resolve an old TODO() about non-v8 build using |JS_| functions.
Remove the JS_ prefix in the SubmitForm() case, is it is not
directly called from fxjs. Remove the JS_ prefix in the
GetFilePath() case, but also provide a JS_ wrapper. Ideally,
all of the JS_ functions would be declared as an interface
in fxjs itself, so this helps keep that consistent.

-- Make SubmitForm() take a span.

Change-Id: I79c3472fa7338112fede7965f5dfa014dad781df