[SkiaPaths] Fix issue that non-Windows cannot load font correctly.
When testing on barcode_test.pdf, error "Could not create FT_Face.\n"
is triggered. This is because |m_FontData| in the temporary CFX_Font
in CFDE_TextOut::DrawString() is not initialized from |pFxFont|, which
later causes CFX_GlyphCache::GetDeviceCache() to fail at building a
valid |m_pTypeface|.

To fix this issue, initialize |m_FontData| in the temporary CFX_Font
structure with |m_FontData| from |pFxFont| to gain access to the font
data.

This CL also removes the unnecessary process of resetting |face| for
the temporary CFX_Font, since it will be destroyed automatically when
CFDE_TextOut::DrawString() finishes.

Bug: pdfium:740
Change-Id: I5e4da4e1928ce21d98393d84f81e58c2a7e479eb