Remove some CPDFXFA_Context usage from top-level fpdfsdk layer.
The CPDF_Document::Extension is intended to keep non-xfa code
blissfully ignorant of the XFA feature bolted onto the side of it.
The cost is a few static_casts<> until the Extension gains a
few more virtual methods to cover these cases.

Use IPDF_Page abstraction in a few places where knowledge of
the underlying CPDFXFA_Page isn't required.

Change-Id: Ie5d484d1c391d9f82947528b4856d7882f8cfad2