Update patch set 3
Patch Set 3:

> Patch Set 1: Commit-Queue+1
> 
> (1 comment)
> 
> Just one nit.
> 
> tsepez@: Do you want to help huyna@ write a test case as a follow-up?

Re: tests - Looking at the stack trace, there is a call to FORM_OnFocus, which means that in addition to the .pdf file, we'd have to create an .evt file.  The stack shows the args to Onfocus as 127,26 which would imply a line in the .evt file:
focus,127,26

but I can't get the test case to reproduce in this manner.

Patch-set: 3
Reviewer: Gerrit User 5495 <5495@fca2d64a-0420-3532-b200-12848c95a47b>