Make CPDF_CMap initialization slightly more efficient.
CPDF_CMapParser is a short lived class that initializes a CPDF_CMap. For
a CPDF_CMap's |m_MixedFourByteLeadingRanges| vector, the parser
initializes it by reading out of it and then appending to it. This
requires CPDF_CMap to have 2 public methods, with the append method
appending one element at a time.

Improve upon this by removing those two methods and instead first let
CPDF_CMapParser hold on to the vector as |m_Ranges|. Since the parser
has direct access to both |m_Ranges| and |m_PendingRanges|, it can move
elements into |m_Ranges| in bulk. When the parser is done, it can use
the newly added CPDF_CMap::SetMixedFourByteLeadingRanges() to move
|m_Ranges| to CPDF_CMap.

Change-Id: I6e6e339296a08a210b925efc278b490a457f8026