Move two files from top-level fpdfsdk/ directory.
These were inconsistently named with regards to the rest of the
files in the directory, and as it turns out, there are better
places for these to live.

CPDF_AnnotContext is a straight move to core/fpdfapi/page.
CPDFSDK_XFAWidget is renamed to CPDFXFA_Widget and moves
to fpdfsdk/fpdfxfa with the other xfa-only files.

The build rules get slightly simpler as a result.

Change-Id: Ie8120b6c037c254ded0ad2c231c680e9435d30f7