PDF a11y: Fill/StrokeColor and Weight APIs
The text fill/stroke color are now exposed. Both APIs takes a boolean
|isRendered| parameter which is set based on the rendering mode. If the
fill color is not rendered, the color value is still returned, but the
|isRendered| parameter is set to false. The same goes for stroke color.

The font-weight is now also exposed as a new parameter of
the FPDFText_GetFontInfo API. If the /FontWeight property is set in the
PDF file, it's used instead of calculating it from the /StemV as we
previously did.

Bug: chromium:985604
Change-Id: Idb414b400feee7cf63d7b87fc50d715997778bf5