Make all CPDFSDK_Annot subclasses contain the word "Annot"
The main confusion is over the use of the term widget to describe
either a CXFA_FFWidget as onwed by a layout item, or a CPDFDSK_Widget
which is a form of annotation. The term "widget annotation" appears in
the PDF spec, so CPDFSDK_WidgetAnnot is suitable for the PDF case.
Then CPDFXFA_WidgetAnnotation is suitable for the corresponding object
in the XFA case.

Change-Id: If7c018b7120179536fcb0cba07f502a988e6f12d