Update patch set 7
Patch Set 7:

(2 comments)

WDYT about just renaming GetXFAWidget() to GetXFAFFWidget() instead? And just be mindful of the fact that there are 2 class hierarchies, one with "FF" in the name and one without.

The entire naming scheme is a bit wonky, so if we really want to make better sense of it all, we are going to have to a lot of renaming. Looking at xfa/fxfa/README.md, I really think CXFA_FFWidget should be CXFA_FFAnnot, and CXFA_FFField should be CXFA_FFWidget. That maps a bit better to Acroform's notion of annots vs widgets. That README even says "CXFA_FFField is the base class for widgets" so CXFA_FFWidget isn't always a widget.

I believe there is a 1:1 relationship between a CXFA_FFWidget and a CXFA_FFWidget, so I think their names should match. If CPDFXFA_Widget becomes CPDFXFA_Annot, then CXFA_FFWidget should become CXFA_FFAnnot. As a result, GetXFAWidget() should become GetXFAAnnot(), but that brings back the naming confusion. We'd want to rename GetXFAAnnot() to GetXFAFFAnnot() after that.

Patch-set: 7
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>