Rename CPDFXFA_Widget classes to CPDFXA_Annot
There is similarity with a CXFA_FFWigdet, which appears in some
of the same files as the CPDFXFA_Widget, and and some GetXFAWidget()
calls which are confusing.  Since CPDFXFA_Widget inherits from
CPDFSDK_Annot, call it an "annot".

Ditto for CPDFXFA_WidgetHandler and CXFA_FFWidgetHandler, which
are similarly distinct.

This may help make a subsequent refactoring saner.

Change-Id: Ie405bbbde6e772b4c4391b9cdcbe170258ae0e74