Fix listbox over-scroll to specified top index
CPDFSDK_AppStream::SetAsListBox() to generate a normal appearance
stream ("N") that includes every list option instead of every option
proceeding the option at the specified top index ("TI"). Instead, the
top indexed option  will be displayed at the top with an offset that can be
reversed if the option is low enough on the list to cause an over-scroll.

The propsed solution assumes that every list option plate has the same
height because the height is determined by the font and font size that
remain unchanged throughout the list options.

Bug: 1386
Change-Id: I746def4bc8aa8eb0f0cfadfac6c65737f2877850