Add tests for listbox form scrolling
Currently, listbox forms do not scroll to the top selected option when
the form is out of focus. Additionally, when a top index ("TI") is
specified for a listbox form, the form will occasionally overscroll.
Tests are added with the current incorrect functionality, and will be
changed in CLs [1] and [2].

Meanwhile update CPDF_FormField::IsItemSelected() to be able to check
for multiple values.

[1] https://pdfium-review.googlesource.com/c/pdfium/+/59190
[2] https://pdfium-review.googlesource.com/c/pdfium/+/59831

Bug: pdfium:1377, pdfium:1386
Change-Id: I67b8bb434d2b8cdcd9a66eca44220a4ad9f86d2b