Add tests for listbox form scrolling
Currently, listbox forms do not scroll to the top selected option when
the form is out of focus. A test is added with the current incorrect
functionality, and will be changed in CL [1].

Another test is added to make sure the listbox does not overscroll when
the top index ("TI") is specified to the last element.

[1] https://pdfium-review.googlesource.com/c/pdfium/+/59190

Bug: pdfium:1377
Change-Id: I67b8bb434d2b8cdcd9a66eca44220a4ad9f86d2b