Scroll to first selected listbox option when out of focus
Implement CPDFSDK_Widget::SetTopVisibleIndex(). Previously, the function
body was not implemeted, and the top visible index was not updated when
the form data was saved.

Update CPDF_FormField::GetTopVisibleIndex() to add a "TI" (top index)
entry to the annotation dictionary if it is missing one so that list
boxes can scroll to the top selected option when the PDF loads.

Update some tests to reflect new correct behavior.

Bug: pdfium:1377
Change-Id: I0884f6ca6d2f20332ca025a55de4103cab1a09ed