New API to retrieve a character's loose bounds
This change adds a new API `FPDFText_GetLooseCharBox` that retrieves the
"loose" bounds of a character.  I.e., the returned bounds will cover the
entire possible extent of the character (given its width).  The height
of the box will always be the same as the font size.  The width will
equal the entire "advance" of the character.

Compare with `FPDFText_GetCharBox`, which returns a "tight" box around
the actual curves that make up the glyph.

Change-Id: I2860f12865b6c81e90cb2621fd8adac1ab5f2530