Load StructTree for Linearized tagged PDFs with /XRefStm
In linearize pdf,we are currently parsing and loading XRefTable, but we
are not loading objects present in XRefStm. Because of that when we
request for struct tree and try to find objects which are in XRefStm it
will return objects are free and so we are not able to get struct tree
for linearizzed pdfs. So Adding method to load XRedStm for linearized
pdf, will able to get struct tree.

Bug: 1335
Change-Id: Ic9650843808f7ed52f6183f306b0b825ec0d57a2