[SkiaPaths]Fix missing italic font style.
Skia API provides reading standard glyphs by their index number and
modifying the glyphs by SkFont structure. When using a substitute font,
its italic angle was not reflected in SkFont during the rendering
process, which caused the italic style to be missing.

Add process to get the italic angle if using a substitute font, and
calculate the |fSkewX| value for SkFont structure.

Bug: pdfium:846
Change-Id: I866a5168e3dcc00da9492a3046254676d5504d56