Update patch set 2
Patch Set 2:

For .in files, we have some rough style guidelines [1], it would be nice to follow them when writing .in files.

For this particular case, I'm not sure how helpful having it as a .in file is vs. having it as a .pdf file, given it has a large amount of binary data. For instance, on https://crbug.com/pdfium/1370, I gave up on handwriting and just checked in a .pdf file, with the "source" being a HTML file. (See comment 2 on the bug.)

[1] https://pdfium.googlesource.com/pdfium/+/master/README.md#files

Patch-set: 2
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>