Update patch set 5
Patch Set 5:

(4 comments)

Using the pdf which is attached to the bug. Script has no support to generate linearized pdf. And we tried several ways to generate linearized tagged pdf whose /StructTreeRoot present in /XrefStm but not working. So, using the same pdf which was attached to the bug.

Patch-set: 5
Reviewer: Gerrit User 7702 <7702@fca2d64a-0420-3532-b200-12848c95a47b>