This is a test for tiling overlapping paths.
When the a colored tiling pattern's opacity value is not 1.0, the
section of overlapped patterns will have a different shade of color.
bug_1286.in is not rendered correctly because the patterns' paths
(borders) are overlapping and SkiaPaths chose to use
kDifference_SkPathOp to subtract the latter path from the previous one,
which result in the overlapping paths to have a lighter color than the
pattern itself. Need to use kUnion_SkPathOp to union the overlapping
paths and skip stroking the path if the path's fill color is the same
as the stroke color to avoid unionizing causing a darker shade of the
overlapping path.

Bug: pdfium:1342
Change-Id: Ideb665ae42bd077f6fe3be364a056286a520a1e6