Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> (1 comment)
> 
> 1. Is there any test coverage? You mentioned this didn't render correctly before in bug_1286.in. Do you need to do anything to indicate it renders correctly now?
> 
> 2. You mentioned that the overlapped part shouldn't be rendered darker if the stroke and fill colors are the same. Can you explain why this is the correct behavior a bit more?

Lei already explained the 1st question so I will answer the 2nd one. This case is to put square tiles right next to each other with no gap. Supposedly, they should just merge into one huge block with no lines or borders on it. AGG backend, adobe and okular all follow this correct behavior.
What happened for SkiaPaths is, for each tile, we fill the area with the fill color and draw the border with the stroke color. So the overlapped border gets painted twice(result in the darker grid lines). To avoid this issue, we can skip drawing the border lines if border lines is the same as the fill color.

Patch-set: 6
Reviewer: Gerrit User 7620 <7620@fca2d64a-0420-3532-b200-12848c95a47b>