Update patch set 6
Patch Set 6:

Just to be clear, I don't have any major concerns with this change, I'm mostly seeking additional information because I don't understand the subject matter. :-) Essentially a soft LGTM from me, pending action items.

> Patch Set 6:
> 
> > Patch Set 6:
> > 
> > (1 comment)
> > 
> > 1. Is there any test coverage? You mentioned this didn't render correctly before in bug_1286.in. Do you need to do anything to indicate it renders correctly now?
> 
> Many test cases in corpus_tests improves with this. But we can't see that here since the bots do not run tests for Skia Paths, since they don't "pass" yet.

It'd be nice to have a better solution for this as the test coverage improves... Would it be feasible to have a list of tests/patterns that are OK to run with Skia paths?

> > 2. You mentioned that the overlapped part shouldn't be rendered darker if the stroke and fill colors are the same. Can you explain why this is the correct behavior a bit more?
> 
> Lei already explained the 1st question so I will answer the 2nd one. This case is to put square tiles right next to each other with no gap. Supposedly, they should just merge into one huge block with no lines or borders on it. AGG backend, adobe and okular all follow this correct behavior.
> What happened for SkiaPaths is, for each tile, we fill the area with the fill color and draw the border with the stroke color. So the overlapped border gets painted twice(result in the darker grid lines). To avoid this issue, we can skip drawing the border lines if border lines is the same as the fill color.

I guess "exact same as the fill color" seems like a fragile condition to me. What if the color is just one level different? Is the result going to look just as wrong as it does now? (Maybe some visual aids would help me understand this better. :-) )

Also, it'd be good to have any additional details near the code somewhere (ideally), or at least the CL description or bug.

Patch-set: 6
Reviewer: Gerrit User 7613 <7613@fca2d64a-0420-3532-b200-12848c95a47b>