Update patch set 1
Patch Set 1:

This is ready for initial review.

Note that I am already unhappy with the naming, simply because the existing code was applying the convention normally reserved for double precision to single precision.  (E.g., in math.h round() is for double and roundf() is for float -- but FXCRT_round() already exists and is for float.)

This initial change set takes the lazy approach and uses FXCRT_dround() for a double.  I'm not thrilled with that.  I think a better change would be for FXCRT_round() to become FXCRT_roundf(), after which FXCRT_round() could be for use with a double.  That would have to trickle through everywhere first, which is much more work - so don't want to start that until some feedback shows agreement that it is indeed the preferred approach.

Patch-set: 1
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>
Reviewer: Gerrit User 7704 <7704@fca2d64a-0420-3532-b200-12848c95a47b>
Work-in-progress: false