Change duplicate test file presubmit check to ignore deletes.
In javascript and pixel tests, there should not be .in and .pdf files
with the same name. There is a presubmit check to enforce this, but it
does not understand file deletions. So if one deletes foo.pdf and add
foo.in in the same CL, the presubmit check will fail. Fix this by
ignoring file deletions in the presubmit check.

Change-Id: I59cc04884f25852e5db749c6242ee4857bbc7a71