Correctly draw text with pattern.
CPDF_RenderStatus::DrawTextPathWithPattern() does the drawing by
creating a new path object in memory and rendering that. However, it
does not take the existing clip path into account. Add
CPDF_ClipPath::CopyClipPath() and use that in DrawTextPathWithPattern()
to copy the existing clip path into the path object that draws the text
path.

Bug: pdfium:632
Change-Id: I6d82697315b4c7537ec2c4136a0689fd5f8a48c7