Fix the TrueType font not rendered correctly issue.
This CL adds an extra check on whether the font is embedded when
determining whether no glyph index is found for a TrueType font
character. For Simple font character, an index of 0xffff is a guaranteed
failure. But for TrueType font (one of the Simple font types), an index
of 0 indicates a freetype failure, which means we should either try to
find the glyph info from the embedded font file or from a fallback font
if there is one.

Bug: pdfium:1388
Change-Id: I8b1935a094f76408e4678723d566914c2bbf0f42