Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> > Patch Set 9:
> > 
> > (5 comments)
> > 
> > Ping. 
> > 
> > I just require the Windows expected PNG, which can be obtained by checking out the linked CL [1] and running pdfium_test on the PDF corresponding to the .in file of this CL:
> > 
> > [1] https://pdfium-review.googlesource.com/c/pdfium/+/59831
> 
> Can do, but one question - do you think we'll end up adding a test to flatten bug_1386.pdf?

I don't believe so? 

We can if you think it's a good idea, but flattening is more of an issue for PDFs with controls that need to be hidden. That wouldn't be an issue with bug_1386.pdf.

Patch-set: 9
Reviewer: Gerrit User 7644 <7644@fca2d64a-0420-3532-b200-12848c95a47b>