Fix build with pdf_is_complete_lib=true and pdf_is_standalone=true
In standalone builds, //:pdfium is jumbo_component, which forwards to a
regular component.  When is_component_build == true, the behavior of
component is to forward to a source_set when sources is not defined, and
forward to a static_library otherwise.  Since //:pdfium has no sources
(but instead pulls in many dependencies as sources), it will be a
source_set, which is not what we want for standalone builds.  Instead,
force it to be a static_library.

BUG=1393
R=thestig

Change-Id: I3c4c090d4bcbfc52406a70bb9abaee00b0172198