Add CPDF_Path::AppendFloatRect().
Make it simpler for callers that already have a CFX_FloatRect to pass
in. Instead of making them pass in the 4 components of the rectangle.
This better aligns with the APIs offered by CFX_PathData, which
CPDF_Path uses. At the same time, rename the CFX_FloatRect version of
CFX_PathData::AppendRect() to avoid overloading.

Change-Id: I0bdd53a34ef8ecaa17a8442299a2a2fcbaa54b17