Add a pixel test for listbox form scrolling behavior
Add tests to ensure that listbox forms do not overscroll when the
specified top index is the last element, or when there is no specified
top index and the last element is the first selected. Add a test to
ensure that listbox forms scroll to a specified top index, when one is
provided, instead of the first selected list element.

Bug: pdfium:1386
Change-Id: I467a27543e51ad13f5dc04b471446982c44216da