Update patch set 15
Patch Set 15:

> Patch Set 12:
> 
> (3 comments)
> 
> Benchmark time for me...

I first changed the affected file to trigger a deliberate crash, so I can identify which PDFs I have that executes this code. Of those, I excluded some of the bigger ones to save time when benchmarking, since benchmarking with testing/tools/safetynet_compare.py using Callgrind is accurate but slow.

That leaves ~500 PDFs for the benchmark set. All of them got slower, as expected. The threhold for performance regressions in SafetyNet is 2%, and 30 PDFs regressed. A bunch are in the 2.x% range, and the worst case is 5.95%.

Patch-set: 15
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>