Update patch set 15
Patch Set 15:

> Patch Set 15:
> 
> > Patch Set 12:
> > 
> > (3 comments)
> > 
> > Benchmark time for me...
> 
> I first changed the affected file to trigger a deliberate crash, so I can identify which PDFs I have that executes this code. Of those, I excluded some of the bigger ones to save time when benchmarking, since benchmarking with testing/tools/safetynet_compare.py using Callgrind is accurate but slow.
> 
> That leaves ~500 PDFs for the benchmark set. All of them got slower, as expected. The threhold for performance regressions in SafetyNet is 2%, and 30 PDFs regressed. A bunch are in the 2.x% range, and the worst case is 5.95%.

The input from the fuzzer has valid |r|, |g|, and |b| values. So at least for that input, clamping is unnecessary, and it is the more expensive part. Without clamping, none of the test PDFs slow down enough to be considered a performance regression.

It is hard to tell if the clamping will be required in other cases, so let's just avoid clamping for now, which fixes the immediately issue, and then let the fuzzer see if it can generate bad values that would trigger the need for clamping.

Patch-set: 15
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>