Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (2 comments)
> 
> > Patch Set 3:
> > 
> > (2 comments)
> > 
> > I'll find some time today to do the performance testing.
> 
> Great, thanks!

In https://pdfium-review.googlesource.com/c/pdfium/+/60030, I mentioned 30 test files regressed for an earlier patch set. For this CL, only 4 regressed, but the comparision isn't apples to apples because this CL builds on top of the prior CL and the performance difference is relative.

Given we probably need this CL to fix the bug, I think the best we can do is to optimize the performance. I have an idea, so I'm benchmarking that.

Patch-set: 4
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>