FPDFAnnot_SetAP API creating a PDF standard compliant annotation
SetAP API exposed from pdfium only sets an appearance stream in
annotation’s dictionary and does not create it as a separate XObject.

Other viewers such as Adobe Reader fail to render such annotation as
these readers expect an XObject to be present with the appearance
stream.

This CL modifies FPDFAnnot_SetAP API to create appearance stream as a
XObject following below steps.
1. Create an indirect PDF stream object.
2. Store appearance stream data in newly created indirect object.
3. Set dictionary values(BBox, Type, SubType) of indirect object.
4. Store a reference to this indirect object in actual annotation
   object's dictionary.

This CL also adds a test case to validate creation of appearance
stream as a separate XObject.

Bug: pdfium:1404
Change-Id: I5054cb15ad8bad6ed55f4b952197249a9cc2dfad