Add a test case for incorrectly rendered Japanese PDFs.
There are non-conforming Japanese language PDFs in wide circulation that
use a Type 2 CIDFont including a CIDToGIDMap where the associated True
Type font program is not embedded. The PDF Reference, Sixth Edition,
specifies that a CIDToGIDMap "may appear only in a Type 2 CIDFont whose
associated True Type font program is embedded in the PDF file"
(Section 5.6, pp 437).

These PDFs render correctly if the illegal CIDToGIDMap is ignored. This
CL adds a test case to illustrate the bug. This CL is based on
https://pdfium-review.googlesource.com/c/pdfium/+/54930

Bug: pdfium:1308
Change-Id: If8b0e0d3f55e8d3216629f769aa5fe62b92ab38b