Add a pixel test for a JPEG2000 decoding issue.
Demonstrate PDFs that contain JPEG2000 images with 4 channels, but have
PDF dictionaries that claim the ColorSpace is RGB, which is only 3
channels, do not render (though they should). As such, the test
expectation is for the desired rendering, and the pixel test is being
added with a suppression. The test PDF actually contains 2 images, one
red and one blue, to help make sure there is not RGB vs. BGR mix up.

Bug: chromium:1012369
Change-Id: If14b07d1bc31586e9e45d01e533c392f8a379837