Fix 64-bit arithmetic in sha2 variants.
SHA-256 has a 8 byte length field, and was being calculated
correctly but clumsily. SHA-384 and SHA-512 have 16 byte length
fields, which were not being calculated properly anyways for
inputs over half a GB (because 32-N bit shifts were used where
64-N bit shifts were required) . Simplify by limiting the length
of the input we support to 2^64 bits for any of these, which
ought to be enough for anyone.

- Rename some macros for consistency and reorder.

Change-Id: I492d0d574177239519ee38c9cbbc2a9beab6967c