Update patch set 7
Patch Set 7:

(5 comments)

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > > > Also, 129 KB is a bit big. If you really can't reduce the file size, put it in the testing corpus instead.
> > > 
> > > Could we embed a smaller font, for example? It doesn't seem like this bug really cares much about the contents of the embedded font, just the presence of an unexpected CIDToGIDMap.
> > 
> > Also, the bytes look pretty repetitive, so even just applying some compression might help a lot here.
> 
> I misread what object 8 0 was (it's the CIDToGIDMap itself). It seems easy enough to trim that down a lot, though, given that we don't have that many characters in this test file.


To make sure the CIDToGIDMap is illegal in the test case, the Type 2 CIDFont has to be not embedded and CIDToGIDMap needs to be a stream object.

kmoon@ Thanks for helping me figure out the right unicode entry for simplifying the test case! I updated the description for the CL, hope it can provide better context this time 😊.

Patch-set: 7
Reviewer: Gerrit User 7620 <7620@fca2d64a-0420-3532-b200-12848c95a47b>