Add public API to support forced colors mode
The CL adds the support for a forced color scheme which is applied
when calling the new progressive render API.
The color scheme is currently applied to only text and path.

For text rendering, the color of the text (fill and stroke) is
changed to the color mentioned in the color scheme.

Similarly colors for fill and stroke can be set for path rendering.
In addition to this there is an option to change the fill paths to
strokes, which is required since with fill paths using a single
color, a graphics with multiple adjacent fill paths will not be
decipherable.

All annotations also follow the above rules, except forms which
are not part of this change, and will be taken up separately.

There also needs to be a special handling for highlight since
the Multiply blend mode leads to the highlights not being visible
in some color schemes, which will be taken up separately.

Bug: 1391
Change-Id: I8b3f43eb0290d11b6a1ee2d5a36dfd1f0c0993e8