Add a pixel test for PDF with illegal CIDToGIDMap stream.
The PDF Reference, Sixth Edition, specifies that a CIDToGIDMap "may
appear only in a Type 2 CIDFont whose associated True Type font
program is embedded in the PDF file" (Section 5.6, pp 437). But some
PDFs, which use Type 2 CIDFont, contain CIDToGIDMap streams even when
the font is not embedded.

This CL adds a pixel test for Type2 CIDFont (with illegal CIDToGIDMap
stream and no CID2UnicodeMap) in addition to a pixel test named
bug_1308.pdf (with illegal CIDToGIDMap stream and CID2UnicodeMap).

Bug: pdfium:1308
Change-Id: I10ae0189ba667c9a7093f8bee1cf77876d611db5