Ignore CID mapping lines which cross first-byte boundaries.
1.4.1 "Cautionary Notes" in Adobe Technical Note #5411 specifies
"when using 'bfrange' operators, care must be taken not to cross
first-byte boundaries for the CID values in a single mapping line.

This CL adds a check for the first bytes of |high| and |low| values
when parsing bfrange. If |high|'s first byte doesn't match |low|'s,
consider current mapping line illegal and ignore it.

Bug:1021762
Change-Id: I4024c216ae6038f3d4270aebc6db86b7c38c0c5c